refactor: 모달 이름, 파일위치 변경
