+++ src/components/ConfirmModal/AddCouponModal.tsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled';
-import { Text } from '../Text/Text';
-import { useModal } from '../../../hooks/useModal';
+import { Text } from '../common/Text/Text';
+import { useModal } from '../../hooks/useModal';
 
-const DeleteCartItemModal = () => {
+const AddCouponModal = () => {
   const { modalDataState, closeModal } = useModal();
 
   const onSubmit = () => {
@@ -14,26 +14,21 @@ const DeleteCartItemModal = () => {
     <ModalWrapper>
       <TextWrapper>
         <Text size="smallest" weight="light">
-          해당 상품을 삭제하시겠습니까?
+          {modalDataState.title}
         </Text>
       </TextWrapper>
       <ButtonWrapper>
-        <CancelButton onClick={closeModal}>
-          <Text size="smallest" weight="light">
-            취소
-          </Text>
-        </CancelButton>
-        <SubmitButton onClick={onSubmit}>
+        <ButtonStyle onClick={onSubmit}>
           <Text size="smallest" weight="normal">
             확인
           </Text>
-        </SubmitButton>
+        </ButtonStyle>
       </ButtonWrapper>
     </ModalWrapper>
   );
 };
 
-export default DeleteCartItemModal;
+export default AddCouponModal;
 
 const ModalWrapper = styled.div`
   display: flex;
@@ -65,7 +60,7 @@ const ButtonWrapper = styled.div`
 `;
 
 const ButtonStyle = styled.button`
-  width: 50%;
+  width: 100%;
   height: 100%;
   display: flex;
   align-items: center;
@@ -75,13 +70,4 @@ const ButtonStyle = styled.button`
   &:hover {
     filter: brightness(0.94);
   }
-`;
-
-const CancelButton = styled(ButtonStyle)`
-  border-radius: 0 0 0 8px;
-  border-right: 1px solid rgba(0, 0, 0, 0.1);
-`;
-
-const SubmitButton = styled(ButtonStyle)`
-  border-radius: 0 0 8px 0;
 `;
