+++ src/components/ConfirmModal/ConfirmModal.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren } from 'react';
 import styled from '@emotion/styled';
 import ReactDOM from 'react-dom';
-import { useModal } from '../../../hooks/useModal';
+import { useModal } from '../../hooks/useModal';
 
 type ModalProps = PropsWithChildren;
